# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediBook Multi-Hospital Doctor Appointment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [Auth]
      summary: Get the current user's app profile (role, etc.)
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateUserProfile
      tags: [Auth]
      summary: Update user profile (role, hospital assignment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserProfile:
      type: object
      required: [id, replitUserId, role]
      properties:
        id:
          type: integer
        replitUserId:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [patient, doctor, hospital_admin, super_admin]
        hospitalId:
          type: [integer, 'null']
        doctorId:
          type: [integer, 'null']
    UpdateUserProfileBody:
      type: object
      properties:
        role:
          type: string
          enum: [patient, doctor, hospital_admin, super_admin]
        hospitalId:
          type: integer
...